% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{add_colors}
\alias{add_colors}
\title{Add random colors to Seurats Object in Misc}
\usage{
add_colors(object, features)
}
\arguments{
\item{object}{a Seurat Object}

\item{features}{Seurat Meta.data columns}
}
\value{
Seurat Object
}
\description{
Add random colors to Seurats Object in Misc
}
\examples{
#don't need
}
