% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{add_image}
\alias{add_image}
\title{Add the image to the seurat object for some software use coord and image information}
\usage{
add_image(seuratObject)
}
\arguments{
\item{seuratObject}{a seurat object with no image s4 obejct}
}
\value{
seuratObject
}
\description{
Add the image to the seurat object for some software use coord and image information
}
\examples{
#
}
