% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{add_loc}
\alias{add_loc}
\title{Title Split rownames into the coordinate position of X and Y.}
\usage{
add_loc(data = NULL, pattern = "a")
}
\arguments{
\item{data}{the data need its rownames including the postion x and y}

\item{pattern}{merge coord x and y use pattern as a split character,default is "a"}
}
\value{
data
}
\description{
Title Split rownames into the coordinate position of X and Y.
}
\examples{
#
}
