% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{adjust_position}
\alias{adjust_position}
\title{Title: Sometimes the slice is not in the correct position on the chip. This
function can make the data point rotate at a specific angle}
\usage{
adjust_position(data = NULL, angle = NULL)
}
\arguments{
\item{data}{the data has x and y column represent coordinate position.}

\item{angle}{the rotate angle, must be a pi number, like pi/2}
}
\value{
data
}
\description{
Title: Sometimes the slice is not in the correct position on the chip. This
function can make the data point rotate at a specific angle
}
\examples{
#
}
