% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cell_in_chip}
\alias{cell_in_chip}
\title{Title: Draw a cell clustering map at the chip position}
\usage{
cell_in_chip(object = NULL, cols = NULL, rotate = NULL, size = 2, ...)
}
\arguments{
\item{object}{the Seurat object}

\item{cols}{the color vector to mapping}

\item{rotate}{the rotate angle, must be a pi number, like pi/2}

\item{size}{the point size}

\item{...}{parameter}
}
\value{
ggplot
}
\description{
Title: Draw a cell clustering map at the chip position
}
\examples{
#
}
