% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_exp_loc}
\alias{get_exp_loc}
\title{Title Retrieves data (feature expression, PCA scores, metrics, etc.) for a
set of cells in a Seurat object and add the coordinate position of X and Y of spatial chip.}
\usage{
get_exp_loc(object, featrues, cells = NULL, slot = "scale.data", ...)
}
\arguments{
\item{object}{the Seurat object}

\item{featrues}{the var want to select}

\item{cells}{the cell want to subset}

\item{slot}{must be one of "counts","data","scale.data"}

\item{...}{inherit}
}
\value{
data
}
\description{
Title Retrieves data (feature expression, PCA scores, metrics, etc.) for a
set of cells in a Seurat object and add the coordinate position of X and Y of spatial chip.
}
\examples{
#
}
