% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{gs}
\alias{gs}
\title{Title the ggplot2 ggsave function changed for the dir path}
\usage{
gs(dir, filename, width, height, dpi)
}
\arguments{
\item{dir}{the dirpath you want to save in}

\item{filename}{the filename you want to name}

\item{width}{the width parameter in ggplot2::ggsave}

\item{height}{the height parameter in ggplot2::ggsave}

\item{dpi}{the dpi parameter in ggplot2::ggsave}
}
\value{
ggplot2 object
}
\description{
Title the ggplot2 ggsave function changed for the dir path
}
\examples{
#
}
