% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{load_spat_env}
\alias{load_spat_env}
\title{Title the packages environment need install or library}
\usage{
load_spat_env()
}
\value{
environment
}
\description{
Title the packages environment need install or library
}
\examples{
#
}
