% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{loci2name}
\alias{loci2name}
\title{Title the St work created by I and qiujiaying,but the workflow don't suit.so make
a "loci2name" function for the object transfrom.}
\usage{
loci2name(object)
}
\arguments{
\item{object}{the seurat spatialObj}
}
\value{
object the
}
\description{
Title the St work created by I and qiujiaying,but the workflow don't suit.so make
a "loci2name" function for the object transfrom.
}
\examples{
#
}
