% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{MIA_analysis}
\alias{MIA_analysis}
\title{Title To integrate the scRNAseq and ST datasets, MIA method can analysis proceeds by first delineating sets of cell type-specific and tissue region-specific genes and then determining whether their overlap is higher (enrichment) or lower (depletion) than expected by chance.}
\usage{
MIA_analysis(sc_object, sp_object, sc_diff, sp_diff)
}
\arguments{
\item{sc_object}{the seurat scRNA object}

\item{sp_object}{the seurat scRNA object or spatial object}

\item{sc_diff}{the different expression result of seurat analysis}

\item{sp_diff}{the different expression result of seurat analysis}
}
\value{
data
}
\description{
Title To integrate the scRNAseq and ST datasets, MIA method can analysis proceeds by first delineating sets of cell type-specific and tissue region-specific genes and then determining whether their overlap is higher (enrichment) or lower (depletion) than expected by chance.
}
\examples{
# d = MIA_analysis(sc_object = stereo_smn, sp_object = stereo, sc_diff = sc_diff, sp_diff = sp_diff)
}
