% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{need_colors}
\alias{need_colors}
\title{Title color palettes created by douyin dadongsheji, I use colors package a R packages}
\usage{
need_colors(colpal, ncol = NULL, style = "china")
}
\arguments{
\item{colpal}{a valid color palette name(color number + id,such as 'two1',can see disco())
or a numberic vector that can select color randomly.}

\item{ncol}{when set the colpal names, ncol set how many color you need}

\item{style}{default parameter, Ancient Chinese style is can't change}
}
\value{
a vector of color code
}
\description{
Title color palettes created by douyin dadongsheji, I use colors package a R packages
}
\examples{
need_colors(9)
need_colors("nine1",8)
need_colors("two2",2)
}
