% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_chip}
\alias{plot_chip}
\title{Title: plot the point in spatial chip}
\usage{
plot_chip(
  data = data,
  features = features,
  size = 2,
  cols = c("white", "#e31a1c"),
  ...
)
}
\arguments{
\item{data}{the dataframe include x and y column represent the location in chip.}

\item{features}{which features you want to display.}

\item{size}{the ggplot2 point size}

\item{cols}{the colors}

\item{...}{parameter}
}
\value{
ggplot2 object
}
\description{
Title: plot the point in spatial chip
}
\examples{
#
}
