% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{random_colors}
\alias{random_colors}
\title{random_colors}
\usage{
random_colors(n)
}
\arguments{
\item{n}{The number of colors.}
}
\value{
A vector of colors.
}
\description{
random_colors
}
\examples{
#don't need
}
