% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{read_spat}
\alias{read_spat}
\title{Title read data from BGI stereo-seq, and creat a seurat object}
\usage{
read_spat(file, pattern = "a", seurat = TRUE, name = "Spatial")
}
\arguments{
\item{file}{the data csv file path}

\item{pattern}{merge coord x and y use pattern as a split character,default is "a"}

\item{seurat}{TRUE/FALSE creat or not creat a seurat object}

\item{name}{the object name}
}
\value{
object
}
\description{
Title read data from BGI stereo-seq, and creat a seurat object
}
\examples{
#
}
