% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{seurat_to_monocle3}
\alias{seurat_to_monocle3}
\title{Title the github website:https://github.com/satijalab/seurat/issues/2833}
\usage{
seurat_to_monocle3(object)
}
\arguments{
\item{object}{the Seurat object and want to analysis in monocle3}
}
\value{
object
}
\description{
Title the github website:https://github.com/satijalab/seurat/issues/2833
}
\examples{
#
}
