% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{SpatialScatterpie}
\alias{SpatialScatterpie}
\title{Title plot Scatterpie for the spotlight result}
\usage{
SpatialScatterpie(
  st_object,
  cell_types,
  slice = "slice1",
  alpha = 1,
  pie_scale = 1
)
}
\arguments{
\item{st_object}{the seurat object with the meta.data including you want to display.}

\item{cell_types}{the names of meta.data colnames which you want to display.}

\item{slice}{slice name, the object name in image slot in seurat object,not the image name.}

\item{alpha}{the point alpha.}

\item{pie_scale}{the pie size.}
}
\value{
ggplot2 object
}
\description{
Title plot Scatterpie for the spotlight result
}
\examples{
# no example
}
