% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{exp_in_chip}
\alias{exp_in_chip}
\title{Title: heatmap on chip}
\usage{
exp_in_chip(
  object,
  featrues,
  cells = NULL,
  slot = "scale.data",
  rotate = NULL,
  cols = NULL,
  ...
)
}
\arguments{
\item{object}{the Seurat object}

\item{featrues}{the gene ID want to display}

\item{cells}{the cell want to subset}

\item{slot}{must be one of "counts","data","scale.data"}

\item{rotate}{the rotate angle, must be a pi number, like pi/2}

\item{cols}{the color vector to mapping}

\item{...}{inherit}
}
\value{
ggplot
}
\description{
Title: heatmap on chip
}
\examples{
#
}
